UPDATED (#): - Mergeo con 12 para incluir todas las últimas modificaciones y correcciones

+++ okticket_connector_hr_expense_sheet/models/hr_expense_sheet.py
@@ -128,8 +128,6 @@ class HrExpenseBatchImporter(Component):
             }
 
             hr_expense_sheet = hr_expense_sheet_obj.search(sheet_domain)
-
-            # try:
             if hr_expense_sheet:  # Update
                 hr_expense_sheet.write(expense_sheet_values)
             else:  # Create
@@ -149,7 +147,8 @@ class HrExpenseBatchImporter(Component):
                 new_sheet_values = hr_expense_sheet_obj.prepare_expense_sheet_values(new_sheet_values)
                 new_sheet_values.update(expense_sheet_values)
                 new_sheet = hr_expense_sheet_obj.create(new_sheet_values)
-                if new_sheet and not new_sheet.okticket_bind_ids:
+                if new_sheet and not new_sheet.okticket_bind_ids \
+                        and self.collection.okticket_exp_sheet_sync:  # Si está activa la sincronización
                     new_sheet.unlink()  # Delete expense sheet. Some error occurs while Okticket sync.
 
     def sanitize_expenses(self, hr_expense_ids):
